[CHORE] extension 분리 (#458)
