+++ Manito/Manito/Screens/Setting/SettingViewController+MailComposeViewControllerDelegate.swift
@@ -1,0 +1,52 @@
+//
+//  SettingViewController+MailComposeViewControllerDelegate.swift
+//  Manito
+//
+//  Created by 이성호 on 2023/05/12.
+//
+
+import MessageUI
+
+// MARK: - MFMailComposeViewControllerDelegate
+extension SettingViewController: MFMailComposeViewControllerDelegate {
+    func sendReportMail() {
+        if MFMailComposeViewController.canSendMail() {
+            let composeVC = MFMailComposeViewController()
+            let aenittoEmail = "aenitto@gmail.com"
+            let messageBody = """
+                              
+                              -----------------------------
+                              
+                              - 문의하는 닉네임: \(String(describing: UserDefaultStorage.nickname))
+                              - 문의 메시지 제목 한줄 요약:
+                              - 문의 날짜: \(Date())
+                              
+                              ------------------------------
+                              
+                              문의 내용을 작성해주세요.
+                              
+                              """
+            
+            composeVC.mailComposeDelegate = self
+            composeVC.setToRecipients([aenittoEmail])
+            composeVC.setSubject("[문의 사항]")
+            composeVC.setMessageBody(messageBody, isHTML: false)
+            
+            self.present(composeVC, animated: true, completion: nil)
+        }
+        else {
+            self.showSendMailErrorAlert()
+        }
+    }
+    
+    private func showSendMailErrorAlert() {
+        let sendMailErrorAlert = UIAlertController(title: "메일 전송 실패", message: "아이폰 이메일 설정을 확인하고 다시 시도해주세요.", preferredStyle: .alert)
+        let confirmAction = UIAlertAction(title: "확인", style: .default)
+        sendMailErrorAlert.addAction(confirmAction)
+        self.present(sendMailErrorAlert, animated: true, completion: nil)
+    }
+    
+    func mailComposeController(_ controller: MFMailComposeViewController, didFinishWith result: MFMailComposeResult, error: Error?) {
+        controller.dismiss(animated: true, completion: nil)
+    }
+}
